HEEDLS-585 Pull out common GetCustomPromptOptions helper function
