+++ DigitalLearningSolutions.Web/ViewModels/TrackingSystem/Delegates/AllDelegates/AllDelegatesViewModelFilterOptions.cs
@@ -47,7 +47,7 @@
                 new FilterViewModel(
                     "JobGroupId",
                     "Job Group",
-                    DelegatesViewModelFilterOptions.GetJobGroupOptions(jobGroups)
+                    DelegatesViewModelFilters.GetJobGroupOptions(jobGroups)
                 ),
                 new FilterViewModel("RegistrationType", "Registration Type", RegistrationTypeOptions)
             };
@@ -56,7 +56,7 @@
                     customPrompt => new FilterViewModel(
                         $"CustomPrompt{customPrompt.CustomPromptNumber}",
                         customPrompt.CustomPromptText,
-                        DelegatesViewModelFilterOptions.GetCustomPromptOptions(customPrompt)
+                        DelegatesViewModelFilters.GetCustomPromptOptions(customPrompt)
                     )
                 )
             );
