+++ DigitalLearningSolutions.Web/ViewModels/TrackingSystem/Delegates/Shared/DelegatesViewModelFilters.cs
@@ -6,9 +6,10 @@
     using DigitalLearningSolutions.Data.Models.User;
     using DigitalLearningSolutions.Web.Helpers;
     using DigitalLearningSolutions.Web.Models.Enums;
+    using DigitalLearningSolutions.Web.ViewModels.Common;
     using DigitalLearningSolutions.Web.ViewModels.Common.SearchablePage;
 
-    public static class DelegatesViewModelFilterOptions
+    public static class DelegatesViewModelFilters
     {
         public static IEnumerable<FilterOptionViewModel> GetJobGroupOptions(
             IEnumerable<(int id, string name)> jobGroups
@@ -51,6 +52,33 @@
                 )
             );
             return options;
+        }
+
+        public static Dictionary<int, string> GetCustomPromptFilters(
+            IEnumerable<CustomFieldViewModel> customFields,
+            IEnumerable<CustomPrompt> promptsWithOptions
+        )
+        {
+            var promptsWithOptionsIds = promptsWithOptions.Select(c => c.CustomPromptNumber);
+            var customFieldsWithOptions =
+                customFields.Where(customField => promptsWithOptionsIds.Contains(customField.CustomFieldId));
+            return customFieldsWithOptions
+                .Select(
+                    customField => new KeyValuePair<int, string>(
+                        customField.CustomFieldId,
+                        GetFilterValueForCustomField(customField)
+                    )
+                ).ToDictionary(x => x.Key, x => x.Value);
+        }
+
+        private static string GetFilterValueForCustomField(CustomFieldViewModel customField)
+        {
+            string filterValueName =
+                CentreCustomPromptHelper.GetDelegateCustomPromptAnswerName(customField.CustomFieldId);
+            string propertyValue = string.IsNullOrEmpty(customField.Answer)
+                ? FilteringHelper.EmptyValue.ToString()
+                : customField.Answer;
+            return FilteringHelper.BuildFilterValueString(filterValueName, filterValueName, propertyValue);
         }
     }
 }
